Move MqttDriver to its own namespace
